chore: move magic tags component under the HFG\Core namespace make sure we have a SSOT for the tags codeinwp/neve-pro-addon#428
