+++ header-footer-grid/Core/Customizer.php
@@ -88,106 +88,7 @@ class Customizer {
 	 * @return mixed
 	 */
 	public function add_dynamic_tags_options( $array ) {
-		$options = [
-			[
-				'label'    => __( 'Single', 'neve' ),
-				'controls' => [
-					'current_single_title'   => [
-						'label' => __( 'Current Single Title', 'neve' ),
-						'type'  => 'string',
-					],
-					'current_single_excerpt' => [
-						'label' => __( 'Current Single Excerpt', 'neve' ),
-						'type'  => 'string',
-					],
-					'current_single_url'     => [
-						'label' => __( 'Current Single URL', 'neve' ),
-						'type'  => 'url',
-					],
-				],
-			],
-			[
-				'label'    => __( 'Archive', 'neve' ),
-				'controls' => [
-					'archive_description' => [
-						'label' => __( 'Archive Description', 'neve' ),
-						'type'  => 'string',
-					],
-					'archive_title'       => [
-						'label' => __( 'Archive Title', 'neve' ),
-						'type'  => 'string',
-					],
-					'archive_url'         => [
-						'label' => __( 'Archive URL', 'neve' ),
-						'type'  => 'url',
-					],
-				],
-			],
-			[
-				'label'    => __( 'Author', 'neve' ),
-				'controls' => [
-					'author_bio'  => [
-						'label' => __( 'Author Bio', 'neve' ),
-						'type'  => 'string',
-					],
-					'author_name' => [
-						'label' => __( 'Author Name', 'neve' ),
-						'type'  => 'string',
-					],
-					'author_url'  => [
-						'label' => __( 'Author URL', 'neve' ),
-						'type'  => 'url',
-					],
-				],
-			],
-			[
-				'label'    => __( 'Global', 'neve' ),
-				'controls' => [
-					'site_title'   => [
-						'label' => __( 'Site Title', 'neve' ),
-						'type'  => 'string',
-					],
-					'site_tagline' => [
-						'label' => __( 'Site Tagline', 'neve' ),
-						'type'  => 'string',
-					],
-					'home_url'     => [
-						'label' => __( 'Home URL', 'neve' ),
-						'type'  => 'url',
-					],
-					'current_year' => [
-						'label' => __( 'Current Year', 'neve' ),
-						'type'  => 'string',
-					],
-				],
-			],
-		];
-
-		if ( class_exists( 'WooCommerce', false ) ) {
-			$options[] = [
-				'label'    => __( 'WooCommerce', 'neve' ),
-				'controls' => [
-					'product_price' => [
-						'label' => __( 'Product Price', 'neve' ),
-						'type'  => 'string',
-					],
-					'product_title' => [
-						'label' => __( 'Product Title', 'neve' ),
-						'type'  => 'string',
-					],
-					'cart_link'     => [
-						'label' => __( 'Cart URL', 'neve' ),
-						'type'  => 'url',
-					],
-					'checkout_link' => [
-						'label' => __( 'Checkout URL', 'neve' ),
-						'type'  => 'url',
-					],
-				],
-			];
-		}
-
-		$array['dynamicTags']['options'] = $options;
+		$array['dynamicTags']['options'] = Magic_Tags::get_instance()->get_options();
 
 		return $array;
 	}
